US3378 One class per file
