+++ Gateway/crds-angular/crds-angular.csproj
@@ -285,7 +285,10 @@
     <Compile Include="Exceptions\NotHeadOfHouseholdException.cs" />
     <Compile Include="Models\Crossroads\AddressDTO.cs" />
     <Compile Include="Models\Crossroads\Childcare\ChildcareDashboardDto.cs" />
+    <Compile Include="Models\Crossroads\Childcare\ChildCareDate.cs" />
+    <Compile Include="Models\Crossroads\Childcare\ChildcareGroup.cs" />
     <Compile Include="Models\Crossroads\Childcare\ChildcareRequestDto.cs" />
+    <Compile Include="Models\Crossroads\Childcare\ChildcareRsvp.cs" />
     <Compile Include="Models\Crossroads\CommunicationDTO.cs" />
     <Compile Include="Models\Crossroads\Congregation.cs" />
     <Compile Include="Models\Crossroads\DonorStatementDTO.cs" />
